@@ -1,47 +1,69 @@
 {
-  "productName": "Eidetica",
-  "name": "eidetica",
-  "description": "Screenshot uploader",
   "author": {
     "name": "Trezy",
     "email": "tre@trezy.com",
     "url": "http://trezy.com"
   },
-  "version": "2.0.0",
+  "description": "Screenshot uploader",
+  "license": "AGPL-3.0",
   "main": "./src/main.js",
+  "name": "eidetica",
+  "productName": "Eidetica",
   "repository": {
     "type": "git",
     "url": "git@github.com:trezy/eidetica.git"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "archiver": "^1.3.0",
-    "electron-compile": "^6.3.0",
-    "electron-config": "^0.2.1",
-    "electron-log": "^2.0.0",
-    "electron-updater": "^1.11.1",
-    "electron-main-notification": "^1.0.1",
-    "react": "^15.4.2",
-    "react-desktop": "^0.2.19",
-    "react-dom": "^15.4.2",
-    "scp2": "^0.5.0",
-    "shorthash": "^0.0.2",
-    "sshpk": "^1.11.0"
-  },
-  "devDependencies": {
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-preset-env": "^1.3.3",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-react": "^6.24.1",
-    "electron-forge": "^2.9.0",
-    "electron-prebuilt-compile": "1.6.5"
   },
   "scripts": {
     "dev": "NODE_ENV='development' yarn run start",
     "start": "electron-forge start",
     "package": "electron-forge package",
+    "postinstall": "electron-rebuild",
+    "make": "electron-forge make",
     "publish": "electron-forge publish",
-    "make": "electron-forge make"
+    "lint": "eslint src --color"
+  },
+  "version": "3.0.0-beta",
+  "dependencies": {
+    "archiver": "2.1.1",
+    "electron-compile": "6.4.2",
+    "electron-config": "1.0.0",
+    "electron-log": "2.2.15",
+    "electron-squirrel-startup": "1.0.0",
+    "electron-updater": "2.21.10",
+    "html-entities": "1.2.1",
+    "prop-types": "^15.6.2",
+    "react": "16.4.2",
+    "react-dom": "16.4.2",
+    "scp2": "0.5.0",
+    "sharp": "0.20.5",
+    "shorthash": "0.0.2",
+    "ssh-config": "1.1.3",
+    "ssh2": "0.6.1",
+    "sshpk": "1.11.0",
+    "untildify": "^3.0.3"
+  },
+  "devDependencies": {
+    "babel-eslint": "8.2.3",
+    "babel-plugin-transform-async-to-generator": "6.24.1",
+    "babel-plugin-transform-class-properties": "6.24.1",
+    "babel-plugin-transform-object-rest-spread": "6.26.0",
+    "babel-preset-env": "1.7.0",
+    "babel-preset-react": "6.24.1",
+    "electron-forge": "5.2.2",
+    "electron-prebuilt-compile": "2.0.2",
+    "electron-rebuild": "^1.8.2",
+    "eslint": "4",
+    "eslint-config-airbnb": "16",
+    "eslint-loader": "^2.0.0",
+    "eslint-plugin-import": "2",
+    "eslint-plugin-jsx-a11y": "6",
+    "eslint-plugin-react": "7"
+  },
+  "build": {
+    "asar": true,
+    "asarUnpack": [
+      "**/node_modules/sharp/**/*"
+    ]
   },
   "config": {
     "forge": {
@@ -50,7 +72,6 @@
           "squirrel"
         ],
         "darwin": [
-          "dmg",
           "zip"
         ],
         "linux": [
@@ -72,13 +93,14 @@
       "electronPackagerConfig": {
         "appBundleId": "eidetica",
         "appCategoryType": "public.app-category.productivity",
-        "appCopyright": "Copyright © 2017 Trezy.com",
+        "appCopyright": "Copyright © 2018 Trezy.com",
         "dir": "./src",
         "icon": "./src/assets/icon",
-        "osxSign": true
+        "osxSign": true,
+        "packageManager": "yarn"
       },
       "electronWinstallerConfig": {
-        "name": ""
+        "name": "blep"
       },
       "electronInstallerDebian": {},
       "electronInstallerRedhat": {},
@@ -87,7 +109,8 @@
         "name": "eidetica"
       },
       "windowsStoreConfig": {
-        "packageName": ""
+        "packageName": "",
+        "name": "blep"
       }
     }
   }
