@@ -1,31 +1,24 @@
+// Module imports
+/* eslint-disable import/no-extraneous-dependencies */
 import {
   app,
-  globalShortcut,
   Menu,
-  nativeImage,
-  Tray
 } from 'electron'
-import notify from 'electron-main-notification'
+/* eslint-enable */
 import fs from 'fs'
+import Config from 'electron-config'
 import log from 'electron-log'
-import path from 'path'
 
-
-
-
-
-import createPane from './modules/common/createPane'
-import handleScreenshot from './modules/common/handleScreenshot'
-import setupApplicationMenu from './modules/common/setupApplicationMenu'
-import setupAutoUpdater from './modules/common/setupAutoUpdater'
-import setupUploadListener from './modules/common/setupUploadListener'
-
-
-
-
-
-// Setup development mode
-let developmentMode = process.env.NODE_ENV === 'development'
+// Component imports
+import {
+  createPane,
+  handleScreenshot,
+  isDevelopmentMode,
+  setupApplicationMenu,
+  setupAutoUpdater,
+  setupUploadListener,
+} from './modules/common'
+import { updateIcon } from './modules/trayIcon'
 
 
 
@@ -37,10 +30,121 @@ setupAutoUpdater()
 
 
 
+// Handle creating/removing shortcuts on Windows when installing/uninstalling.
+if (require('electron-squirrel-startup')) { // eslint-disable-line global-require
+  app.quit()
+}
+
+
+
+
+
 new class App {
+  /***************************************************************************\
+    Local Properties
+  \***************************************************************************/
+
+  fillPad = 4
+
+  screenshotQueue = new Set([])
+
+  size = 45
+
+
+
+
+
+  /***************************************************************************\
+    Local Methods
+  \***************************************************************************/
+
+  initialize = async () => {
+    log.info('Initializing app')
+
+    // Set up the default items for all context menus
+    setupApplicationMenu()
+
+    // Get the screenshot folder
+    this.screenshotFolder = app.getPath('desktop')
+
+    // By default, close events shouldn't actually close the app
+    app.shouldQuit = false
+
+    // Prevent the dock icon from being displayed in production
+    if (!isDevelopmentMode()) {
+      app.dock.hide()
+    }
+
+    // Preload the pane in a hidden browser window
+    // app.pane = createPane()
+
+    // Create the system tray icon
+    await updateIcon()
+
+    const menuOptions = [
+      {
+        label: 'Preferences...',
+        click: createPane,
+      },
+      {
+        label: 'Quit',
+        click: () => {
+          app.shouldQuit = true
+          app.quit()
+        },
+      },
+    ]
+
+    if (isDevelopmentMode()) {
+      menuOptions.push({
+        label: 'Restart',
+        click: () => {
+          app.shouldQuit = true
+          app.relaunch()
+          app.quit()
+        },
+      })
+    }
+
+    app.tray.setToolTip(app.getName())
+    app.tray.setContextMenu(Menu.buildFromTemplate(menuOptions))
+
+    app.on('will-quit', event => (!app.shouldQuit ? event.preventDefault() : null))
+
+    // Start listening for screenshots to be taken
+    this.startScreenshotListener()
+
+    setupUploadListener()
+
+    log.info('App initialized')
+  }
+
+  startScreenshotListener = () => {
+    // MacOS
+    fs.watch(this.screenshotFolder, (type, filename) => {
+      if (/^Screen Shot.*\.png$/gi.exec(filename) && !this.screenshotQueue.has(filename)) {
+        this.screenshotQueue.add(filename)
+        handleScreenshot(filename)
+      }
+    })
+
+    // Windows
+    // globalShortcut.register('PrintScreen', () => {
+    //   log.info('PrintScreen pressed')
+    // })
+  }
+
+
+
+
+
+  /***************************************************************************\
+    Public Methods
+  \***************************************************************************/
+
   constructor () {
     // Setup default configs
-    new (require('electron-config'))({
+    new Config({
       defaults: {
         // App settings
         autoUpdate: true,
@@ -56,90 +160,10 @@ new class App {
         port: '',
         url: '',
         username: '',
-      }
+      },
     })
-
-    // Bind any functions that will always require the app as context
-    this.initialize = this.initialize.bind(this)
-    this.showPreferencesPane = this.showPreferencesPane.bind(this)
 
     // Spin up application
     app.on('ready', this.initialize)
-  }
-
-  initialize () {
-    log.info('Initializing app')
-
-    let trayIconFilename = 'tray-iconTemplate.png'
-    let trayIconPath
-
-    // Set up the default items for all context menus
-    setupApplicationMenu()
-
-    // Get the screenshot folder
-    this.screenshotFolder = app.getPath('desktop')
-
-    // Track whether or not we should exit or just hide the preferences pane
-    this.shouldQuitApp = false
-
-    // Prevent the dock icon from being displayed
-    app.dock.hide()
-
-    // Preload the pane in a hidden browser window
-    this.pane = createPane()
-
-    // Attach the pane to the app object to make it easier to access from inside the pane itself
-    app.pane = this.pane
-
-    // Figure out where the tray icon lives
-    if (process.env.NODE_ENV === 'development') {
-      trayIconPath = path.resolve(__dirname, 'assets', trayIconFilename)
-    } else {
-      trayIconPath = path.resolve(process.resourcesPath, 'app', 'src', 'assets', trayIconFilename)
-    }
-
-    // Generate the tray icon as a native image
-    this.trayIcon = nativeImage.createFromPath(trayIconPath)
-
-    // setting the tray icon as a template image means macOS will handle changing it from white to black for us
-    this.trayIcon.setTemplateImage(true)
-
-    this.tray = new Tray(this.trayIcon)
-    this.tray.setToolTip(app.getName())
-    this.tray.setContextMenu(Menu.buildFromTemplate([
-      {
-        label: 'Preferences...',
-        click: this.showPreferencesPane
-      },
-      {
-        label: 'Quit',
-        click: app.quit
-      }
-    ]))
-
-    // Start listening for screenshots to be taken
-    this.startScreenshotListener()
-
-    setupUploadListener()
-
-    log.info('App initialized')
-  }
-
-  showPreferencesPane () {
-    this.pane.loadPane('index')
-  }
-
-  startScreenshotListener () {
-    // MacOS
-    fs.watch(this.screenshotFolder, (type, filename) => {
-      if (filename.indexOf('Screen Shot') === 0) {
-        handleScreenshot(filename)
-      }
-    })
-
-    // Windows
-  //  globalShortcut.register('PrintScreen', () => {
-  //    log.info('PrintScreen pressed')
-  //  })
   }
 }
