@@ -1,8 +1,8 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 import {
   app,
-  BrowserWindow
+  BrowserWindow,
 } from 'electron'
-import log from 'electron-log'
 import path from 'path'
 import url from 'url'
 
@@ -10,37 +10,55 @@ import url from 'url'
 
 
 
-import setupUploadListener from './setupUploadListener'
+import {
+  isDevelopmentMode,
+  setupUploadListener,
+} from '.'
 
 
 
 
 
-module.exports = function () {
-  let pane = new BrowserWindow({
-    frame: false,
-    height: 500,
-    resizable: false,
+const createPane = () => {
+  const pane = new BrowserWindow({
+    backgroundColor: 'blue',
+    devTools: isDevelopmentMode(),
+    frame: true,
+    fullscreenable: false,
+    height: 400,
+    minHeight: 400,
+    minWidth: 600,
+    movable: true,
     show: false,
-    transparent: false,
+    titleBarStyle: 'hiddenInset',
     useContentSize: true,
-    width: 500,
+    width: 600,
   })
 
-//  pane.openDevTools()
-
-  pane.loadPane = function (pane) {
-    this.loadURL(url.format({
-      pathname: path.join(__dirname, '..', '..', 'panes', `${pane}.html`),
+  pane.loadPane = (paneName) => {
+    pane.loadURL(url.format({
+      pathname: path.resolve(__dirname, '..', '..', 'panes', `${paneName}.html`),
       protocol: 'file',
       slashes: true,
     }))
   }
 
-  pane.on('blur', pane.hide)
+  pane.loadPane('index')
+
+  if (isDevelopmentMode()) {
+    pane.webContents.openDevTools({ mode: 'undocked' })
+  }
+
+  pane.on('beforeunload', () => delete app.pane)
   pane.on('ready-to-show', pane.show)
   pane.on('shortcut-reset', setupUploadListener)
   pane.on('shortcut-updated', setupUploadListener)
 
-  return pane
+  app.pane = pane
 }
+
+
+
+
+
+export { createPane }
