@@ -1,34 +1,34 @@
-import { app } from 'electron'
 import fs from 'fs'
 import log from 'electron-log'
-import path from 'path'
-import shorthash from 'shorthash'
 
 
 
 
 
-import generateTempFilepath from './generateTempFilepath'
+import { generateTempFilepath } from '.'
 
 
 
 
 
-module.exports = function (text) {
-  return new Promise((resolve, reject) => {
-    log.info('Creating txt file')
+const createTextFile = text => new Promise((resolve, reject) => {
+  log.info('Creating txt file')
 
-    let filepath = generateTempFilepath('txt')
+  const filepath = generateTempFilepath('txt')
 
-    fs.writeFile(filepath, text, error => {
-      if (error) {
-        reject(error)
+  fs.writeFile(filepath, text, error => {
+    if (error) {
+      reject(error)
+    } else {
+      log.info('Done.')
 
-      } else {
-        log.info('Done.')
-
-        resolve(filepath)
-      }
-    })
+      resolve(filepath)
+    }
   })
-}
+})
+
+
+
+
+
+export { createTextFile }
