@@ -1,27 +1,47 @@
-import Config from 'electron-config'
+import { readFileSync } from 'fs'
+import { homedir } from 'os'
+import { resolve } from 'path'
+import ElectronConfig from 'electron-config'
+import SSHConfig from 'ssh-config'
+import untildify from 'untildify'
 
 
 
 
 
-let config = new Config
+const config = new ElectronConfig
 
 
 
 
 
-module.exports = function () {
-  let ret = {
-    host: config.get('host'),
-    port: config.get('port') || 22,
-    username: config.get('username'),
-    path: config.get('path'),
+const getSCPConfig = () => {
+  const { host } = config.get()
+  const sshConfig = SSHConfig.parse(readFileSync(resolve(homedir(), '.ssh', 'config'), 'utf8'))
+  const hostConfig = sshConfig.compute(host)
+
+  const scpConfig = {}
+
+  if (hostConfig.IdentityFile) {
+    scpConfig.privateKey = readFileSync(resolve(untildify(hostConfig.IdentityFile[0])))
   }
-  let password = config.get('password')
+
+  scpConfig.host = hostConfig.HostName || host
+  scpConfig.path = config.get('path')
+  scpConfig.port = hostConfig.Port || 22
+  scpConfig.username = hostConfig.User || config.get('username')
+
+  const password = config.get('password')
 
   if (password) {
-    ret.password = password
+    scpConfig.password = password
   }
 
-  return ret
+  return scpConfig
 }
+
+
+
+
+
+export { getSCPConfig }
