@@ -1,35 +1,35 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { app } from 'electron'
 import Config from 'electron-config'
 import fs from 'fs'
 import log from 'electron-log'
 import path from 'path'
-import shorthash from 'shorthash'
 
 
 
 
 
-import copyFile from './copyFile'
-import generateShortlink from './generateShortlink'
-import generateTempFilepath from './generateTempFilepath'
-import uploadFile from './uploadFile'
+import {
+  copyFile,
+  generateTempFilepath,
+  uploadFile,
+} from '.'
 
 
 
 
 
-let config = new Config
+const config = new Config
 
 
 
 
 
-module.exports = function (filename) {
+const handleScreenshot = async filename => {
   log.info('Handling screenshot')
 
-  let filepath = path.resolve(app.getPath('desktop'), filename)
-  let tempFilepath = generateTempFilepath(path.extname(filename), filename)
-  let shortlink = generateShortlink(path.basename(tempFilepath))
+  const filepath = path.resolve(app.getPath('desktop'), filename)
+  const tempFilepath = generateTempFilepath(path.extname(filename), filename)
 
   try {
     fs.readFileSync(filepath)
@@ -38,13 +38,16 @@ module.exports = function (filename) {
     return
   }
 
-  log.info(`A screenshot was captured at ${filename.replace('Screen Shot ', '').replace('.png', '')}`)
+  await copyFile(filepath, tempFilepath)
+  await uploadFile(tempFilepath)
 
-  copyFile(filepath, tempFilepath)
-  .then(uploadFile)
-  .then(() => {
-    if (config.get('deleteAfterUpload')) {
-      fs.unlinkSync(filepath)
-    }
-  })
+  if (config.get('deleteAfterUpload')) {
+    fs.unlinkSync(filepath)
+  }
 }
+
+
+
+
+
+export { handleScreenshot }
